# Speedup observable by avoiding outer products