--- a/test/test_aqua.jl
+++ b/test/test_aqua.jl
@@ -1,5 +1,7 @@
 @testitem "Aqua" tags=[:aqua] begin
 using Aqua
+using QuantumOpticsBase
+using QuantumSavory
 
 if get(ENV,"JET_TEST","")=="true"
 # JET generates new methods with ambiguities
@@ -9,10 +11,10 @@ else
 
 Aqua.test_all(QuantumSavory,
     ambiguities=(QuantumSavory; recursive=false),
-    piracies=(; treat_as_own=[QuantumSavory.Symbolic]),
+    piracies=(; treat_as_own=[QuantumSavory.Symbolic, QuantumOpticsBase.Ket, QuantumOpticsBase.Operator]),
     stale_deps=(; ignore=[:NetworkLayout]) # needed by package extension but not a condition of its loading
 )
 
-@test length(Aqua.Piracy.hunt(QuantumSavory)) == 6
+@test length(Aqua.Piracy.hunt(QuantumSavory)) == 7
 end
 end