--- a/test/test_circuitzoo_purification.jl
+++ b/test/test_circuitzoo_purification.jl
@@ -154,7 +154,7 @@ end
                     initialize!(r[3:4], noisy_pair)
                     initialize!(r[5:6], noisy_pair)
                     if Purify3to1(leaveout1, leaveout2)(r[1], r[2], r[3], r[5], r[4], r[6])==true
-                        @test_broken observable(r[1:2], projector(bell)) ≈ 0.0 # This is a probabilistic test. It has a small chance of triggering
+                        @test_broken observable(r[1:2], projector(bell)) ≈ 0.0 atol=1e-5 # This is a probabilistic test. It has a small chance of triggering
                     end
                 end
             end
@@ -253,7 +253,7 @@ end
                     ma = Purify3to1Node(leaveout1, leaveout2)(r[1], r[3], r[5])
                     mb = Purify3to1Node(leaveout1, leaveout2)(r[2], r[4], r[6])
                     if ma == mb
-                        @test_broken observable(r[1:2], projector(bell)) ≈ 0.0 # This is a probabilistic test. It has a small chance of triggering
+                        @test_broken observable(r[1:2], projector(bell)) ≈ 0.0 atol=1e-5 # This is a probabilistic test. It has a small chance of triggering
                     end
                 end
             end
@@ -300,7 +300,7 @@ end
             initialize!(r[(2*i-1):(2*i)], noisy_pair)
         end
         if PurifyStringent()(r[1], r[2], r[3:2:25]..., r[4:2:26]...) == true
-            @test_broken observable(r[1:2], projector(bell)) ≈ 0.0 # This is a probabilistic test. It has a small chance of triggering
+            @test_broken observable(r[1:2], projector(bell)) ≈ 0.0 atol=1e-5 # This is a probabilistic test. It has a small chance of triggering
         end
     end
 end
@@ -344,7 +344,7 @@ end
             initialize!(r[(2*i-1):(2*i)], noisy_pair)
         end
         if PurifyExpedient()(r[1], r[2], r[3:2:21]..., r[4:2:22]...) == true
-            @test_broken observable(r[1:2], projector(bell)) ≈ 0.0 # This is a probabilistic test. It has a small chance of triggering
+            @test_broken observable(r[1:2], projector(bell)) ≈ 0.0 atol=1e-5 # This is a probabilistic test. It has a small chance of triggering
         end
     end
 end