Fixes #216 (at least partially -- it is a speadup of more than 10x for 5-qubit states)

```
BenchmarkTools.Trial: 10000 samples with 1 evaluation per sample.
 Range (min … max):   5.831 μs …   4.401 ms  ┊ GC (min … max):  0.00% … 99.13%
 Time  (median):      8.336 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   13.733 μs ± 147.468 μs  ┊ GC (mean ± σ):  36.98% ±  3.43%

      ▁▃▄▄▄▄▇▄▆▆▇▆▆█▅▅▄▄▂▁▁
  ▁▂▅███████████████████████▇▆▆▄▅▄▃▃▃▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▁▁▁▁▁▁ ▄
  5.83 μs         Histogram: frequency by time           15 μs <

 Memory estimate: 28.28 KiB, allocs estimate: 160.
```

```
BenchmarkTools.Trial: 10000 samples with 1 evaluation per sample.
 Range (min … max):  400.000 ns …  11.031 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     481.000 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   522.458 ns ± 197.006 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

    ▃█▇ ▁
  ▂▄███▃█▇▅▆▆▂▅▅▅▅▅▅▂▄▄▄▄▄▂▄▃▃▃▃▂▃▃▃▃▃▁▃▃▃▃▃▁▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ ▃
  400 ns           Histogram: frequency by time          912 ns <

 Memory estimate: 2.23 KiB, allocs estimate: 17.
```